+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Modeler3D:
   - Set points while drawing can be undone with CTRL + Z and redone with CTRL + Y
   - The movement of objects can be undone with CTRL + Z and redone with CTRL + Y
+  - Add edit function for fillColor and stroke color of drawn polygons and polylines.
 - overviewMap:
   - Adds New parameter `renderIn3d` for overviewMap. Enables the overviewMap in 3d mode.
   - Adds New parameter `boxBorderColor` for overviewMap. Enables to configure the border color of mini overview map box.